requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
pillow>=10.0.0
schedule>=1.2.0
watchdog>=3.0.0
chromadb>=0.4.22
ollama>=0.1.0
numpy>=1.24.0
pyautogui>=0.9.54


